CI try to fix check_local_time integ test
Jobs
https://github.com/ScaleComputing/HyperCoreAnsibleCollection/actions/runs/4675127388/jobs/8279941573#step:8:52
and
https://github.com/ScaleComputing/HyperCoreAnsibleCollection/actions/runs/4675127388/jobs/8279941761#step:8:52
failed

Extra debug print added. It is not clear what were input values,
and what went wrong.

Tested interval is also changed, this should make test more reliable.

Downside is test cannot work 2 hours in a day - there is extra assert
just to fail early when this happens. It should not happen during
scheduled CI jobs.

Removed the extra "" in asserts. We were checking string is not empty,
but we need to let assert evaluate string.

Signed-off-by: Justin Cinkelj <justin.cinkelj@xlab.si>