CI do not assume test ansible controller has configured timezone
Failed CI job
https://github.com/ScaleComputing/HyperCoreAnsibleCollection/actions/runs/4696220997/jobs/8326124886#step:8:49

The timezone from https://ipapi.co/timezone is correct one, it
is based on public IP address/geolocation.
But out test ansible controller is not configured with timezone.
UTC is just OK for servers.

We can convert unix epoch to local time with python and use pytz for that.
pytz is used only for testing.

The timezone from ipapi.co is not needed anymore.

Signed-off-by: Justin Cinkelj <justin.cinkelj@xlab.si>