Cleanup unneeded condition
The version_update_status_info uses undocumented API endpoint,
it is hard to know what is returned. I believe we wanted to have
`==COMPLETE`, not `==COMPLETED`.

I believe whole condition is there only to prevent starting
second upgrade before the first one is finished -
`.update_status != "IN PROGRESS"` is sufficient.

A corner case would be if `.update_status == SOME_ERROR` or similar.
In this case `!= "IN PROGRESS"` would still allow starting an upgrade.
The `==COMPLETE` is again not needed - better to remove it,
it just adds to confusion.

Signed-off-by: Justin Cinkelj <justin.cinkelj@xlab.si>